Be able to encode 64bit value

+++ src/qq_mem/src/utils.cc
@@ -243,7 +243,7 @@ std::string str_qq_search_reply(const qq::SearchReply &reply) {
 // TODO: expand more then 5 so we do not frequently resize? but
 // C++ will expand capacity automatically. We are not actually
 // resizing memory.
-int varint_expand_and_encode(uint32_t value, std::string *buf, const off_t offset) {
+int varint_expand_and_encode(uint64_t value, std::string *buf, const off_t offset) {
   if ( buf->size() - offset < 5 ) {
     buf->resize(offset + 5);
   }
@@ -254,7 +254,7 @@ int varint_expand_and_encode(uint32_t value, std::string *buf, const off_t offse
 // buf must be at least 5 bytes long, which is the size required
 // to store a full unsigned int
 // It returns the number of bytes stored in buf
-int varint_encode(uint32_t value, std::string *buf, const off_t offset) {
+int varint_encode(uint64_t value, std::string *buf, const off_t offset) {
   int i = 0;
   // inv: value has what left to be encoded
   //      buf[offset, offset+i) has encoded bytes
