+++ src/qq_mem/src/utils.h
@@ -180,8 +180,8 @@ std::string format_with_commas(T value)
 
 
 std::string str_qq_search_reply(const qq::SearchReply &reply);
-int varint_expand_and_encode(uint32_t value, std::string *buf, const off_t offset);
-int varint_encode(uint32_t value, std::string *buf, off_t offset);
+int varint_expand_and_encode(uint64_t value, std::string *buf, const off_t offset);
+int varint_encode(uint64_t value, std::string *buf, off_t offset);
 // int varint_decode(const std::string &buf, off_t offset, uint32_t *value);
 // int varint_decode_chars(const char *buf, const off_t offset, uint32_t *value);
